+++ packages/twenty-front/src/modules/object-metadata/utils/__tests__/formatFieldMetadataItemInput.test.ts
@@ -1,59 +1,33 @@
+import { FieldMetadataItemOption } from '@/object-metadata/types/FieldMetadataItem';
 import { FieldMetadataType } from '~/generated-metadata/graphql';
 
-import {
-  formatFieldMetadataItemInput,
-  getOptionValueFromLabel,
-} from '../formatFieldMetadataItemInput';
-
-describe('getOptionValueFromLabel', () => {
-  it('should return the option value from the label', () => {
-    const label = 'Example Label';
-    const expected = 'EXAMPLE_LABEL';
-
-    const result = getOptionValueFromLabel(label);
-
-    expect(result).toEqual(expected);
-  });
-
-  it('should handle labels with accents', () => {
-    const label = 'Éxàmplè Làbèl';
-    const expected = 'EXAMPLE_LABEL';
-
-    const result = getOptionValueFromLabel(label);
-
-    expect(result).toEqual(expected);
-  });
-
-  it('should handle labels with special characters', () => {
-    const label = 'Example!@#$%^&*() Label';
-    const expected = 'EXAMPLE_LABEL';
-
-    const result = getOptionValueFromLabel(label);
-
-    expect(result).toEqual(expected);
-  });
-
-  it('should handle labels with emojis', () => {
-    const label = '📱 Example Label';
-    const expected = 'EXAMPLE_LABEL';
-
-    const result = getOptionValueFromLabel(label);
-
-    expect(result).toEqual(expected);
-  });
-});
+import { formatFieldMetadataItemInput } from '../formatFieldMetadataItemInput';
 
 describe('formatFieldMetadataItemInput', () => {
   it('should format the field metadata item input correctly', () => {
+    const options: FieldMetadataItemOption[] = [
+      {
+        id: '1',
+        label: 'Option 1',
+        color: 'red' as const,
+        position: 0,
+        value: 'OPTION_1',
+      },
+      {
+        id: '2',
+        label: 'Option 2',
+        color: 'blue' as const,
+        position: 1,
+        value: 'OPTION_2',
+      },
+    ];
     const input = {
+      defaultValue: "'OPTION_1'",
       label: 'Example Label',
       icon: 'example-icon',
       type: FieldMetadataType.Select,
       description: 'Example description',
-      options: [
-        { id: '1', label: 'Option 1', color: 'red' as const, isDefault: true },
-        { id: '2', label: 'Option 2', color: 'blue' as const },
-      ],
+      options,
     };
 
     const expected = {
@@ -61,22 +35,7 @@ describe('formatFieldMetadataItemInput', () => {
       icon: 'example-icon',
       label: 'Example Label',
       name: 'exampleLabel',
-      options: [
-        {
-          id: '1',
-          label: 'Option 1',
-          color: 'red',
-          position: 0,
-          value: 'OPTION_1',
-        },
-        {
-          id: '2',
-          label: 'Option 2',
-          color: 'blue',
-          position: 1,
-          value: 'OPTION_2',
-        },
-      ],
+      options,
       defaultValue: "'OPTION_1'",
     };
 
@@ -108,15 +67,29 @@ describe('formatFieldMetadataItemInput', () => {
   });
 
   it('should format the field metadata item multi select input correctly', () => {
+    const options: FieldMetadataItemOption[] = [
+      {
+        id: '1',
+        label: 'Option 1',
+        color: 'red' as const,
+        position: 0,
+        value: 'OPTION_1',
+      },
+      {
+        id: '2',
+        label: 'Option 2',
+        color: 'blue' as const,
+        position: 1,
+        value: 'OPTION_2',
+      },
+    ];
     const input = {
+      defaultValue: ["'OPTION_1'", "'OPTION_2'"],
       label: 'Example Label',
       icon: 'example-icon',
       type: FieldMetadataType.MultiSelect,
       description: 'Example description',
-      options: [
-        { id: '1', label: 'Option 1', color: 'red' as const, isDefault: true },
-        { id: '2', label: 'Option 2', color: 'blue' as const, isDefault: true },
-      ],
+      options,
     };
 
     const expected = {
@@ -124,22 +97,7 @@ describe('formatFieldMetadataItemInput', () => {
       icon: 'example-icon',
       label: 'Example Label',
       name: 'exampleLabel',
-      options: [
-        {
-          id: '1',
-          label: 'Option 1',
-          color: 'red',
-          position: 0,
-          value: 'OPTION_1',
-        },
-        {
-          id: '2',
-          label: 'Option 2',
-          color: 'blue',
-          position: 1,
-          value: 'OPTION_2',
-        },
-      ],
+      options,
       defaultValue: ["'OPTION_1'", "'OPTION_2'"],
     };
